hawkbit: move send_feedback_internal to common
Will be used by the cancel API.
